+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+* Added a diff checking function in response to the feedback I received on [Reddit](https://www.reddit.com/r/FlutterDev/comments/1ejrf4b/package_for_easy_undo_and_redo/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button). Thanks!
+* Added nowIndex method to FileStateManager.
+* Added skipNextPush method to FileStateManager.
+
 ## 1.0.0
 
 * Added now and getStack method to FileStateManager class.